changes in : profile.html, profile.js. Created profileEdit.html and started the connection with the database
profile.html-profile.js : changed the code for the inputs in the
informations and used javascript in the both of them for the connection
with the server. In the the profile.js was created a var person with
elements to be connected with the server and called in the profile.html.
Missing the implement of the data in the profile.js.
In the profileEdit.html the user will be able to write in the text and
press the "save" which will send the new data to the server. Missing
part of the connection.